Corrected 102-fibonacci.c and updated README.md

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -3,7 +3,7 @@
 int main()
 {
 	int i;
-	unsigned long long p1, p2, curr;
+	unsigned long p1, p2, curr;
 
 	p1 = 1, p2 = 2;
 	printf("1, 2, ");
@@ -12,10 +12,10 @@ int main()
 		curr = p1 + p2;
 		if (i == 48)
 		{
-			printf("%llu", curr);
+			printf("%lu", curr);
 			continue;
 		}
-		printf("%llu, ", curr);
+		printf("%lu, ", curr);
 		p1 = p2;
 		p2 = curr;
 	}
